fix: problems of instrumentation-anthropic ts
